+++ internal/aghnet/etchostscontainer.go
@@ -1,8 +1,4 @@
-// Package util contains various utilities.
-//
-// TODO(a.garipov): Such packages are widely considered an antipattern.  Remove
-// this when we refactor our project structure.
-package util
+package aghnet
 
 import (
 	"bufio"
@@ -16,7 +12,6 @@ import (
 	"strings"
 	"sync"
 
-	"github.com/AdguardTeam/AdGuardHome/internal/aghnet"
 	"github.com/AdguardTeam/AdGuardHome/internal/aghos"
 	"github.com/AdguardTeam/golibs/log"
 	"github.com/fsnotify/fsnotify"
@@ -25,8 +20,11 @@ import (
 
 type onChangedT func()
 
-// AutoHosts - automatic DNS records
-type AutoHosts struct {
+// EtcHostsContainer - automatic DNS records
+//
+// TODO(e.burkov): Move the logic under interface.  Refactor.  Probably remove
+// the resolving logic.
+type EtcHostsContainer struct {
 	// lock protects table and tableReverse.
 	lock sync.RWMutex
 	// table is the host-to-IPs map.
@@ -47,63 +45,67 @@ type AutoHosts struct {
 }
 
 // SetOnChanged - set callback function that will be called when the data is changed
-func (a *AutoHosts) SetOnChanged(onChanged onChangedT) {
-	a.onChanged = onChanged
+func (ehc *EtcHostsContainer) SetOnChanged(onChanged onChangedT) {
+	ehc.onChanged = onChanged
 }
 
 // Notify other modules
-func (a *AutoHosts) notify() {
-	if a.onChanged == nil {
+func (ehc *EtcHostsContainer) notify() {
+	if ehc.onChanged == nil {
 		return
 	}
-	a.onChanged()
+	ehc.onChanged()
 }
 
 // Init - initialize
 // hostsFn: Override default name for the hosts-file (optional)
-func (a *AutoHosts) Init(hostsFn string) {
-	a.table = make(map[string][]net.IP)
-	a.onlyWritesChan = make(chan fsnotify.Event, 2)
+func (ehc *EtcHostsContainer) Init(hostsFn string) {
+	ehc.table = make(map[string][]net.IP)
+	ehc.onlyWritesChan = make(chan fsnotify.Event, 2)
 
-	a.hostsFn = "/etc/hosts"
+	ehc.hostsFn = "/etc/hosts"
 	if runtime.GOOS == "windows" {
-		a.hostsFn = os.ExpandEnv("$SystemRoot\\system32\\drivers\\etc\\hosts")
+		ehc.hostsFn = os.ExpandEnv("$SystemRoot\\system32\\drivers\\etc\\hosts")
 	}
 	if len(hostsFn) != 0 {
-		a.hostsFn = hostsFn
+		ehc.hostsFn = hostsFn
 	}
 
 	if aghos.IsOpenWrt() {
 		// OpenWrt: "/tmp/hosts/dhcp.cfg01411c".
-		a.hostsDirs = append(a.hostsDirs, "/tmp/hosts")
+		ehc.hostsDirs = append(ehc.hostsDirs, "/tmp/hosts")
 	}
 
 	// Load hosts initially
-	a.updateHosts()
+	ehc.updateHosts()
 
 	var err error
-	a.watcher, err = fsnotify.NewWatcher()
+	ehc.watcher, err = fsnotify.NewWatcher()
 	if err != nil {
-		log.Error("autohosts: %s", err)
+		log.Error("etchostscontainer: %s", err)
 	}
 }
 
 // Start - start module
-func (a *AutoHosts) Start() {
-	log.Debug("Start AutoHosts module")
+func (ehc *EtcHostsContainer) Start() {
+	if ehc == nil {
+		return
+	}
 
-	a.updateHosts()
+	log.Debug("Start etchostscontainer module")
 
-	if a.watcher != nil {
-		go a.watcherLoop()
+	ehc.updateHosts()
 
-		err := a.watcher.Add(a.hostsFn)
+	if ehc.watcher != nil {
+		go ehc.watcherLoop()
+
+		err := ehc.watcher.Add(ehc.hostsFn)
 		if err != nil {
-			log.Error("Error while initializing watcher for a file %s: %s", a.hostsFn, err)
+			log.Error("Error while initializing watcher for a file %s: %s", ehc.hostsFn, err)
 		}
 
-		for _, dir := range a.hostsDirs {
-			err = a.watcher.Add(dir)
+		for _, dir := range ehc.hostsDirs {
+			err = ehc.watcher.Add(dir)
 			if err != nil {
 				log.Error("Error while initializing watcher for a directory %s: %s", dir, err)
 			}
@@ -112,67 +114,71 @@ func (a *AutoHosts) Start() {
 }
 
 // Close - close module
-func (a *AutoHosts) Close() {
-	if a.watcher != nil {
-		_ = a.watcher.Close()
+func (ehc *EtcHostsContainer) Close() {
+	if ehc == nil {
+		return
 	}
-	close(a.onlyWritesChan)
+
+	if ehc.watcher != nil {
+		_ = ehc.watcher.Close()
+	}
+	close(ehc.onlyWritesChan)
 }
 
 // Process returns the list of IP addresses for the hostname or nil if nothing
 // found.
-func (a *AutoHosts) Process(host string, qtype uint16) []net.IP {
+func (ehc *EtcHostsContainer) Process(host string, qtype uint16) []net.IP {
 	if qtype == dns.TypePTR {
 		return nil
 	}
 
 	var ipsCopy []net.IP
-	a.lock.RLock()
-	defer a.lock.RUnlock()
+	ehc.lock.RLock()
+	defer ehc.lock.RUnlock()
 
-	if ips, ok := a.table[host]; ok {
+	if ips, ok := ehc.table[host]; ok {
 		ipsCopy = make([]net.IP, len(ips))
 		copy(ipsCopy, ips)
 	}
 
-	log.Debug("autohosts: answer: %s -> %v", host, ipsCopy)
+	log.Debug("etchostscontainer: answer: %s -> %v", host, ipsCopy)
 	return ipsCopy
 }
 
 // ProcessReverse processes a PTR request.  It returns nil if nothing is found.
-func (a *AutoHosts) ProcessReverse(addr string, qtype uint16) (hosts []string) {
+func (ehc *EtcHostsContainer) ProcessReverse(addr string, qtype uint16) (hosts []string) {
 	if qtype != dns.TypePTR {
 		return nil
 	}
 
-	ipReal := aghnet.UnreverseAddr(addr)
+	ipReal := UnreverseAddr(addr)
 	if ipReal == nil {
 		return nil
 	}
 
 	ipStr := ipReal.String()
 
-	a.lock.RLock()
-	defer a.lock.RUnlock()
+	ehc.lock.RLock()
+	defer ehc.lock.RUnlock()
 
-	hosts = a.tableReverse[ipStr]
+	hosts = ehc.tableReverse[ipStr]
 
 	if len(hosts) == 0 {
 		return nil // not found
 	}
 
-	log.Debug("autohosts: reverse-lookup: %s -> %s", addr, hosts)
+	log.Debug("etchostscontainer: reverse-lookup: %s -> %s", addr, hosts)
 
 	return hosts
 }
 
 // List returns an IP-to-hostnames table.  It is safe for concurrent use.
-func (a *AutoHosts) List() (ipToHosts map[string][]string) {
-	a.lock.RLock()
-	defer a.lock.RUnlock()
+func (ehc *EtcHostsContainer) List() (ipToHosts map[string][]string) {
+	ehc.lock.RLock()
+	defer ehc.lock.RUnlock()
 
-	ipToHosts = make(map[string][]string, len(a.tableReverse))
-	for k, v := range a.tableReverse {
+	ipToHosts = make(map[string][]string, len(ehc.tableReverse))
+	for k, v := range ehc.tableReverse {
 		ipToHosts[k] = v
 	}
 
@@ -180,7 +186,7 @@ func (a *AutoHosts) List() (ipToHosts map[string][]string) {
 }
 
 // update table
-func (a *AutoHosts) updateTable(table map[string][]net.IP, host string, ipAddr net.IP) {
+func (ehc *EtcHostsContainer) updateTable(table map[string][]net.IP, host string, ipAddr net.IP) {
 	ips, ok := table[host]
 	if ok {
 		for _, ip := range ips {
@@ -199,17 +205,17 @@ func (a *AutoHosts) updateTable(table map[string][]net.IP, host string, ipAddr n
 		ok = true
 	}
 	if ok {
-		log.Debug("autohosts: added %s -> %s", ipAddr, host)
+		log.Debug("etchostscontainer: added %s -> %s", ipAddr, host)
 	}
 }
 
 // updateTableRev updates the reverse address table.
-func (a *AutoHosts) updateTableRev(tableRev map[string][]string, newHost string, ipAddr net.IP) {
+func (ehc *EtcHostsContainer) updateTableRev(tableRev map[string][]string, newHost string, ipAddr net.IP) {
 	ipStr := ipAddr.String()
 	hosts, ok := tableRev[ipStr]
 	if !ok {
 		tableRev[ipStr] = []string{newHost}
-		log.Debug("autohosts: added reverse-address %s -> %s", ipStr, newHost)
+		log.Debug("etchostscontainer: added reverse-address %s -> %s", ipStr, newHost)
 
 		return
 	}
@@ -221,20 +227,20 @@ func (a *AutoHosts) updateTableRev(tableRev map[string][]string, newHost string,
 	}
 
 	tableRev[ipStr] = append(tableRev[ipStr], newHost)
-	log.Debug("autohosts: added reverse-address %s -> %s", ipStr, newHost)
+	log.Debug("etchostscontainer: added reverse-address %s -> %s", ipStr, newHost)
 }
 
 // Read IP-hostname pairs from file
 // Multiple hostnames per line (per one IP) is supported.
-func (a *AutoHosts) load(table map[string][]net.IP, tableRev map[string][]string, fn string) {
+func (ehc *EtcHostsContainer) load(table map[string][]net.IP, tableRev map[string][]string, fn string) {
 	f, err := os.Open(fn)
 	if err != nil {
-		log.Error("autohosts: %s", err)
+		log.Error("etchostscontainer: %s", err)
 		return
 	}
 	defer f.Close()
 	r := bufio.NewReader(f)
-	log.Debug("autohosts: loading hosts from file %s", fn)
+	log.Debug("etchostscontainer: loading hosts from file %s", fn)
 
 	for done := false; !done; {
 		var line string
@@ -242,7 +248,7 @@ func (a *AutoHosts) load(table map[string][]net.IP, tableRev map[string][]string
 		if err == io.EOF {
 			done = true
 		} else if err != nil {
-			log.Error("autohosts: %s", err)
+			log.Error("etchostscontainer: %s", err)
 
 			return
 		}
@@ -276,8 +282,8 @@ func (a *AutoHosts) load(table map[string][]net.IP, tableRev map[string][]string
 				host = host[:sharp]
 			}
 
-			a.updateTable(table, host, ip)
-			a.updateTableRev(tableRev, host, ip)
+			ehc.updateTable(table, host, ip)
+			ehc.updateTableRev(tableRev, host, ip)
 			if sharp >= 0 {
 				// Skip the comments again.
 				break
@@ -287,20 +293,20 @@ func (a *AutoHosts) load(table map[string][]net.IP, tableRev map[string][]string
 }
 
 // onlyWrites is a filter for (*fsnotify.Watcher).Events.
-func (a *AutoHosts) onlyWrites() {
-	for event := range a.watcher.Events {
+func (ehc *EtcHostsContainer) onlyWrites() {
+	for event := range ehc.watcher.Events {
 		if event.Op&fsnotify.Write == fsnotify.Write {
-			a.onlyWritesChan <- event
+			ehc.onlyWritesChan <- event
 		}
 	}
 }
 
 // Receive notifications from fsnotify package
-func (a *AutoHosts) watcherLoop() {
-	go a.onlyWrites()
+func (ehc *EtcHostsContainer) watcherLoop() {
+	go ehc.onlyWrites()
 	for {
 		select {
-		case event, ok := <-a.onlyWritesChan:
+		case event, ok := <-ehc.onlyWritesChan:
 			if !ok {
 				return
 			}
@@ -310,7 +316,7 @@ func (a *AutoHosts) watcherLoop() {
 			repeat := true
 			for repeat {
 				select {
-				case _, ok = <-a.onlyWritesChan:
+				case _, ok = <-ehc.onlyWritesChan:
 					repeat = ok
 				default:
 					repeat = false
@@ -318,48 +324,48 @@ func (a *AutoHosts) watcherLoop() {
 			}
 
 			if event.Op&fsnotify.Write == fsnotify.Write {
-				log.Debug("autohosts: modified: %s", event.Name)
-				a.updateHosts()
+				log.Debug("etchostscontainer: modified: %s", event.Name)
+				ehc.updateHosts()
 			}
 
-		case err, ok := <-a.watcher.Errors:
+		case err, ok := <-ehc.watcher.Errors:
 			if !ok {
 				return
 			}
-			log.Error("autohosts: %s", err)
+			log.Error("etchostscontainer: %s", err)
 		}
 	}
 }
 
 // updateHosts - loads system hosts
-func (a *AutoHosts) updateHosts() {
+func (ehc *EtcHostsContainer) updateHosts() {
 	table := make(map[string][]net.IP)
 	tableRev := make(map[string][]string)
 
-	a.load(table, tableRev, a.hostsFn)
+	ehc.load(table, tableRev, ehc.hostsFn)
 
-	for _, dir := range a.hostsDirs {
+	for _, dir := range ehc.hostsDirs {
 		fis, err := ioutil.ReadDir(dir)
 		if err != nil {
 			if !errors.Is(err, os.ErrNotExist) {
-				log.Error("autohosts: Opening directory: %q: %s", dir, err)
+				log.Error("etchostscontainer: Opening directory: %q: %s", dir, err)
 			}
 
 			continue
 		}
 
 		for _, fi := range fis {
-			a.load(table, tableRev, filepath.Join(dir, fi.Name()))
+			ehc.load(table, tableRev, filepath.Join(dir, fi.Name()))
 		}
 	}
 
 	func() {
-		a.lock.Lock()
-		defer a.lock.Unlock()
+		ehc.lock.Lock()
+		defer ehc.lock.Unlock()
 
-		a.table = table
-		a.tableReverse = tableRev
+		ehc.table = table
+		ehc.tableReverse = tableRev
 	}()
 
-	a.notify()
+	ehc.notify()
 }